+++ ordersmanagement/BL/BL.Util/Exceptions/ErrorMessage.cs
@@ -6,7 +6,7 @@ namespace BL.Util.Exceptions
 {
     public static class ErrorMessage
     {
-        public static (string Name, string Desciption) OrderNotFound(int orderId) =>
+        public static (string Name, string Desciption) OrderNotFound(long orderId) =>
             ($"Order {orderId}", $"Order not found");
     }
 }
